+++ packages/chameleon-loader/src/cml-compile/runtime/app.js
@@ -1,3 +1,4 @@
 import _cml_ from 'chameleon-runtime';
-
-exports.default = _cml_.createApp(exports.default).getOptions()
+module.exports = function(){
+  _cml_.createApp(exports.default).getOptions();
+}