+++ packages/chameleon-loader/src/cml-compile/runtime/web/app.js
@@ -1,0 +1,2 @@
+import _cml_ from 'chameleon-runtime';
+exports.default = _cml_.createApp(exports.default).getOptions()