+++ packages/chameleon-loader/src/cml-compile/runtime/web/page.js
@@ -1,0 +1,3 @@
+import _cml_ from 'chameleon-runtime';
+
+exports.default = _cml_.createPage(exports.default).getOptions()