+++ packages/chameleon-loader/src/loader.js
@@ -7,7 +7,7 @@ const loaderUtils = require('loader-utils')
 const normalize = require('./utils/normalize')
 const componentNormalizerPath = normalize.lib('runtime/component-normalizer')
 const fs = require('fs');
-const getRunTimeSnippet = require('./cml-compile/runtime/index.js');
+const {getVueRunTimeSnippet} = require('./cml-compile/runtime/index.js');
 
 var compileTemplate = require('chameleon-template-parse');
 
@@ -97,7 +97,8 @@ module.exports = function (content) {
   const isWxmlComponent = extName === '.wxml';
   const isAxmlComponent = extName === '.axml';
   const isSwanComponent = extName === '.swan';
-  const isMiniAppRawComponent = isWxmlComponent ||  isAxmlComponent || isSwanComponent;
+  const isQmlComponent = extName === '.qml';
+  const isMiniAppRawComponent = isWxmlComponent ||  isAxmlComponent || isSwanComponent || isQmlComponent;
   if(!isMiniAppRawComponent) {
     //处理script cml-type为json的内容
     content = cmlUtils.deleteScript({content, cmlType: 'json'});
@@ -164,7 +165,7 @@ module.exports = function (content) {
     qq: 'qml'
   }
   //小程序模板后缀正则
-  const miniTplExtReg = /(\.wxml|\.axml)$/;
+  const miniTplExtReg = /(\.wxml|\.axml|\.swan|\.qml)$/;
   const miniCmlReg = /(\.cml|\.wx\.cml|\.alipay\.cml|\.qq\.cml|\.baidu\.cml)$/;
 
   if(isMiniAppRawComponent) {
@@ -208,7 +209,6 @@ module.exports = function (content) {
 
   // 引用微信小程序组件处理
   function miniAppRawComponentHandler() {
-    
     if((cmlType === 'wx' && extName === '.wxml') || (cmlType === 'alipay' && extName === '.axml') || (cmlType === 'baidu' && extName === '.swan') || (cmlType === 'qq' && extName === '.qml')) {
       //生成json文件
       let jsonFile = filePath.replace(miniTplExtReg,'.json');
@@ -223,8 +223,10 @@ module.exports = function (content) {
       miniAppScript.addMiniAppScript(self,filePath,context,cmlType)
       var styleString = getWxmlRequest('styles');
       var scriptString = getWxmlRequest('script');
+      var entryBasePath = entryPath.replace(miniTplExtReg, '');
       output += `var __cml__style = ${styleString};\n`
-      output += `var __cml__script = ${scriptString};\n`
+      output += `var __cml__script = ${scriptString};\n
+      __CML__GLOBAL.__CMLCOMPONNETS__['${entryBasePath}'] = __cml__script;\n`
 
       //采用分离的方式，入口js会放到static/js下，需要再生成入口js去require该js
       var jsFileName = entryPath.replace(miniTplExtReg, '.js');
@@ -240,7 +242,6 @@ module.exports = function (content) {
     npmComponents.forEach(item=>{
       componentDeps.push(item.filePath);
     })
-
     let newJsonObj = jsonHandler(self, jsonObject, cmlType, componentDeps) || {};
     newJsonObj.usingComponents = newJsonObj.usingComponents || {};
     let usingComponents ={} ;
@@ -310,9 +311,9 @@ module.exports = function (content) {
       var scriptRequireString = script.src
         ? getRequireForImport('script', script)
         : getRequire('script', script)
-
-      output += `var __cml__script = ${scriptRequireString};\n`
-
+      var entryBasePath = entryPath.replace(miniCmlReg, '');
+      output += `var __cml__script = ${scriptRequireString};\n
+      __CML__GLOBAL.__CMLCOMPONNETS__['${entryBasePath}'] = __cml__script;\n`
 
       //采用分离的方式，入口js会放到static/js下，需要再生成入口js去require该js
       var jsFileName = entryPath.replace(miniCmlReg, '.js');
@@ -349,7 +350,7 @@ module.exports = function (content) {
       entryContent += "__CML__GLOBAL.Page = Page;\n"
     }
     entryContent += `require('${relativePath}/static/js/common.js')\n`;
-    entryContent += `require('${relativePath}/static/js/${jsFileName}')\n`;
+    entryContent += `require('${relativePath}/static/js/${jsFileName}')()\n`;
     
      
     self.emitFile(jsFileName, entryContent);
@@ -427,7 +428,7 @@ module.exports = function (content) {
 
     function handleVueScript() {
       let { defineComponets, componetsStr } = getComponents();
-      let runtimeSnippet = getRunTimeSnippet(cmlType, type);
+      let runtimeSnippet = getVueRunTimeSnippet(cmlType, type);
       let scriptCode = getScriptCode(self, cmlType, scriptContent, media, options.check);
       return `
         ${defineComponets}